samples/evaluator-optimizer-due-diligence/requirements.txt
langchain-azure-ai>=1.1.0
langgraph>=0.4.0
langchain-core>=0.3.0
langchain-openai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
azure-ai-projects>=2.0.0
azure-identity>=1.15.0
